Refactor ProfessorCheckinButton and fixed a typo
